fix(tests): increase minimum Pest requirement to 3.0 and remove RefreshDatabase from TestCase.php
Workflow tests are still failing due to 'undefined properties' in Laravel 10 and Laravel 11 prefer-lowest, in addition to the 'There is already an active transaction' error in Laravel 11 prefer-lowest. RefreshDatabase has been replaced with a call to ->artisan('migrate:reset'); in teardown(). This change is designed to hopefully fix both issues.